Split PlayerMgmt controller into player/admin
