+++ src/Nether.Web/Features/PlayerManagement/Controllers/PlayerAdminController.cs
@@ -19,139 +19,21 @@ using Nether.Web.Features.PlayerManagement.Models.PlayerManagement;
 namespace Nether.Web.Features.PlayerManagement
 {
     /// <summary>
-    /// Player management
+    /// Player management admin API
     /// </summary>
-    public class PlayerManagementController : Controller
+    [Authorize(Roles = RoleNames.Admin)]
+    [Route("admin")]
+    public class PlayerAdminController : Controller
     {
         private readonly IPlayerManagementStore _store;
         private readonly ILogger _logger;
 
-        public PlayerManagementController(IPlayerManagementStore store, ILogger<PlayerManagementController> logger)
+        public PlayerAdminController(IPlayerManagementStore store, ILogger<PlayerAdminController> logger)
         {
             _store = store;
             _logger = logger;
         }
 
-        // Implementation of the player API
-        // There are two views:
-        //  1. By Player
-        //  2. Admininstration
-
-        /// <summary>
-        /// Gets the player information from currently logged in user
-        /// </summary>
-        /// <returns></returns>
-        [ProducesResponseType((int)HttpStatusCode.OK, Type = typeof(PlayerGetResponseModel))]
-        [ProducesResponseType((int)HttpStatusCode.NotFound)]
-        [Authorize(Roles = RoleNames.Player)]
-        [HttpGet("player")]
-        public async Task<ActionResult> GetCurrentPlayer()
-        {
-            string gamertag = User.GetGamerTag();
-            if (string.IsNullOrWhiteSpace(gamertag))
-            {
-                return NotFound();
-            }
-
-            // Call data store
-            var player = await _store.GetPlayerDetailsByGamertagAsync(gamertag);
-            if (player == null)
-                return NotFound();
-
-            // Return result
-            return Ok(PlayerGetResponseModel.FromPlayer(player));
-        }
-
-        /// <summary>
-        /// Gets the player information from currently logged in user
-        /// </summary>
-        /// <returns></returns>
-        [ProducesResponseType((int)HttpStatusCode.NoContent)]
-        [ProducesResponseType((int)HttpStatusCode.NotFound)]
-        [Authorize(Roles = RoleNames.Player)]
-        [HttpDelete("player")]
-        public async Task<ActionResult> DeleteCurrentPlayer()
-        {
-            string gamertag = User.GetGamerTag();
-            if (string.IsNullOrWhiteSpace(gamertag))
-            {
-                return NotFound();
-            }
-
-            // Call data store
-            await _store.DeletePlayerDetailsAsync(gamertag);
-
-            // Return result
-            return NoContent();
-        }
-
-        /// <summary>
-        /// Gets the extended player information from currently logged in user
-        /// </summary>
-        /// <returns></returns>
-        [ProducesResponseType((int)HttpStatusCode.OK, Type = typeof(PlayerExtendedGetResponseModel))]
-        [ProducesResponseType((int)HttpStatusCode.NotFound)]
-        [Authorize(Roles = RoleNames.Player)]
-        [HttpGet("playerextended")]
-        public async Task<ActionResult> GetCurrentPlayerExtended()
-        {
-            string userId = User.GetId();
-
-            // Call data store
-            var player = await _store.GetPlayerDetailsExtendedAsync(userId);
-            if (player == null)
-                return NotFound();
-
-            // Return result
-            return Ok(PlayerExtendedGetResponseModel.FromPlayer(player));
-        }
-
-        /// <summary>
-        ///  Updates information about the current player
-        /// </summary>
-        /// <param name="player">Player data</param>
-        /// <returns></returns>
-        [ProducesResponseType((int)HttpStatusCode.NoContent)]
-        [Authorize(Roles = RoleNames.Player)]
-        [HttpPut("player")]
-        public async Task<ActionResult> PutCurrentPlayer([FromBody]PlayerPutRequestModel player)
-        {
-            string userId = User.GetId();
-
-            // TODO - need to prevent modifying gamertag
-
-            // Update player
-            await _store.SavePlayerAsync(
-                new Player { UserId = userId, Gamertag = player.Gamertag, Country = player.Country, CustomTag = player.CustomTag });
-
-            // Return result
-            return NoContent();
-        }
-
-        /// <summary>
-        /// Updates extended (e.g. JSON) information about the current player
-        /// </summary>
-        /// <param name="player">Player data</param>
-        /// <returns></returns>
-        [ProducesResponseType((int)HttpStatusCode.NoContent)]
-        [Authorize(Roles = RoleNames.Player)]
-        [HttpPut("playerextended")]
-        public async Task<ActionResult> PutCurrentPlayerExtended([FromBody]PlayerExtendedPutRequestModel player)
-        {
-            string userId = User.GetId();
-
-            // Update extended player information
-            // Update player
-            await _store.SavePlayerExtendedAsync(
-                new PlayerExtended { UserId = userId, Gamertag = player.Gamertag, ExtendedInformation = player.ExtendedInformation });
-
-            // Return result
-            return NoContent();
-        }
-
-        ////////////////////////////////////
-        // Implementation of the Admin API
-        ///////////////////////////////////
 
         /// <summary>
         /// Creates or updates information about a player. You have to be an administrator to perform this action.
@@ -160,7 +42,6 @@ namespace Nether.Web.Features.PlayerManagement
         /// <returns></returns>
         [ProducesResponseType((int)HttpStatusCode.Created)]
         [ProducesResponseType((int)HttpStatusCode.BadRequest)]
-        [Authorize(Roles = RoleNames.Admin)]
         [HttpPost("players")]
         public async Task<ActionResult> Post([FromBody]PlayerPostRequestModel newPlayer)
         {
@@ -190,7 +71,6 @@ namespace Nether.Web.Features.PlayerManagement
         /// <returns></returns>
         [ProducesResponseType((int)HttpStatusCode.Created)]
         [ProducesResponseType((int)HttpStatusCode.BadRequest)]
-        [Authorize(Roles = RoleNames.Admin)]
         [HttpPost("playersextended")]
         public async Task<ActionResult> PostExtended([FromBody]PlayerExtendedPostRequestModel Player)
         {
@@ -219,7 +99,6 @@ namespace Nether.Web.Features.PlayerManagement
         /// <returns>Player information</returns>
         [ProducesResponseType((int)HttpStatusCode.OK, Type = typeof(PlayerGetResponseModel))]
         [ProducesResponseType((int)HttpStatusCode.NotFound)]
-        [Authorize(Roles = RoleNames.Admin)]
         [HttpGet("players/{gamertag}", Name = nameof(GetPlayer))]
         public async Task<ActionResult> GetPlayer(string gamertag)
         {
@@ -239,7 +118,6 @@ namespace Nether.Web.Features.PlayerManagement
         /// <returns>Player extended information</returns>
         [ProducesResponseType((int)HttpStatusCode.OK, Type = typeof(PlayerExtendedGetResponseModel))]
         [ProducesResponseType((int)HttpStatusCode.NotFound)]
-        [Authorize(Roles = RoleNames.Admin)]
         [HttpGet("playersextended/{gamertag}")]
         public async Task<ActionResult> GetPlayerExtended(string gamertag)
         {
@@ -257,7 +135,6 @@ namespace Nether.Web.Features.PlayerManagement
         /// </summary>
         /// <returns></returns>
         [ProducesResponseType((int)HttpStatusCode.OK, Type = typeof(PlayerListGetResponseModel))]
-        [Authorize(Roles = RoleNames.Admin)]
         [HttpGet("players")]
         public async Task<ActionResult> GetPlayers()
         {
@@ -274,30 +151,18 @@ namespace Nether.Web.Features.PlayerManagement
             return Ok(resultModel);
         }
 
-        /// <summary>
-        /// Gets the list of groups current player belongs to.
-        /// </summary>
-        /// <returns></returns>
-        [ProducesResponseType((int)HttpStatusCode.OK, Type = typeof(GroupListResponseModel))]
-        [Authorize(Roles = RoleNames.PlayerOrAdmin)]
-        [HttpGet("player/groups")]
-        public async Task<ActionResult> GetPlayerGroups()
-        {
-            return await GetPlayerGroups(User.GetGamerTag());
-        }
 
         /// <summary>
         /// Gets the list of group a player belongs to.
         /// </summary>
-        /// <param name="gamerTag">Player's gamertag.</param>
+        /// <param name="gamertag">Player's gamertag.</param>
         /// <returns></returns>
         [ProducesResponseType((int)HttpStatusCode.OK, Type = typeof(GroupListResponseModel))]
-        [Authorize(Roles = RoleNames.Admin)]
-        [HttpGet("players/{gamerTag}/groups")]
-        public async Task<ActionResult> GetPlayerGroups(string gamerTag)
+        [HttpGet("players/{gamertag}/groups")]
+        public async Task<ActionResult> GetPlayerGroups(string gamertag)
         {
             // Call data store
-            var groups = await _store.GetPlayersGroupsAsync(gamerTag);
+            var groups = await _store.GetPlayersGroupsAsync(gamertag);
 
             // Return result
             return Ok(GroupListResponseModel.FromGroups(groups));
@@ -307,14 +172,13 @@ namespace Nether.Web.Features.PlayerManagement
         /// <summary>
         /// Adds player to a group.
         /// </summary>
-        /// <param name="playerName">Player's gamer tag</param>
+        /// <param name="gamertag">Player's gamer tag</param>
         /// <param name="groupName">Group name.</param>
         /// <returns></returns>
         [ProducesResponseType((int)HttpStatusCode.NoContent)]
         [ProducesResponseType((int)HttpStatusCode.BadRequest)]
-        [Authorize(Roles = RoleNames.Admin)]
-        [HttpPut("players/{playerName}/groups/{groupName}")]
-        public async Task<ActionResult> AddPlayerToGroup(string playerName, string groupName)
+        [HttpPut("players/{gamertag}/groups/{groupName}")]
+        public async Task<ActionResult> AddPlayerToGroup(string gamertag, string groupName)
         {
             Group group = await _store.GetGroupDetailsAsync(groupName);
             if (group == null)
@@ -323,10 +187,10 @@ namespace Nether.Web.Features.PlayerManagement
                 return BadRequest();
             }
 
-            Player player = await _store.GetPlayerDetailsByGamertagAsync(playerName);
+            Player player = await _store.GetPlayerDetailsByGamertagAsync(gamertag);
             if (player == null)
             {
-                _logger.LogWarning("player '{0}' not found", playerName);
+                _logger.LogWarning("player '{0}' not found", gamertag);
                 return BadRequest();
             }
 
@@ -335,32 +199,19 @@ namespace Nether.Web.Features.PlayerManagement
             return NoContent();
         }
 
-        /// <summary>
-        /// Adds currently logged in player to a group.
-        /// </summary>
-        /// <param name="groupName">Group name.</param>
-        /// <returns></returns>
-        [ProducesResponseType((int)HttpStatusCode.NoContent)]
-        [ProducesResponseType((int)HttpStatusCode.BadRequest)]
-        [Authorize(Roles = RoleNames.PlayerOrAdmin)]
-        [HttpPut("player/groups/{groupName}")]
-        public async Task<ActionResult> AddCurrentPlayerToGroup(string groupName)
-        {
-            return await AddPlayerToGroup(User.GetGamerTag(), groupName);
-        }
+
 
         /// <summary>
         /// Removes a player from a group.
         /// </summary>
         /// <param name="groupName">Group name</param>
-        /// <param name="playerName">Player name</param>
+        /// <param name="gamertag">Player name</param>
         /// <returns></returns>
         [ProducesResponseType((int)HttpStatusCode.NoContent)]
-        [Authorize(Roles = RoleNames.Admin)]
-        [HttpDelete("groups/{groupName}/players/{playerName}")]
-        public async Task<ActionResult> DeletePlayerFromGroup(string groupName, string playerName)
+        [HttpDelete("groups/{groupName}/players/{gamertag}")]
+        public async Task<ActionResult> DeletePlayerFromGroup(string groupName, string gamertag)
         {
-            Player player = await _store.GetPlayerDetailsByGamertagAsync(playerName);
+            Player player = await _store.GetPlayerDetailsByGamertagAsync(gamertag);
             Group group = await _store.GetGroupDetailsAsync(groupName);
 
             await _store.RemovePlayerFromGroupAsync(group, player);
@@ -376,7 +227,6 @@ namespace Nether.Web.Features.PlayerManagement
         /// <param name="group">Group object</param>
         /// <returns></returns>
         [ProducesResponseType((int)HttpStatusCode.Created)]
-        [Authorize(Roles = RoleNames.PlayerOrAdmin)]
         [HttpPost("groups")]
         public async Task<ActionResult> PostGroup([FromBody]GroupPostRequestModel group)
         {
@@ -412,7 +262,6 @@ namespace Nether.Web.Features.PlayerManagement
         /// </summary>
         /// <returns></returns>
         [ProducesResponseType((int)HttpStatusCode.OK, Type = typeof(GroupListResponseModel))]
-        [Authorize(Roles = RoleNames.Admin)]
         [HttpGet("groups")]
         public async Task<ActionResult> GetGroupsAsync()
         {
@@ -429,7 +278,6 @@ namespace Nether.Web.Features.PlayerManagement
         /// <param name="groupName"></param>
         /// <returns></returns>
         [ProducesResponseType((int)HttpStatusCode.OK, Type = typeof(GroupGetResponseModel))]
-        [Authorize(Roles = RoleNames.Admin)]
         [HttpGet("groups/{groupName}", Name = nameof(GetGroup))]
         public async Task<ActionResult> GetGroup(string groupName)
         {
@@ -456,7 +304,6 @@ namespace Nether.Web.Features.PlayerManagement
         /// <param name="groupName">Name of the group</param>
         /// <returns></returns>
         [ProducesResponseType((int)HttpStatusCode.OK, Type = typeof(GroupMemberResponseModel))]
-        [Authorize(Roles = RoleNames.PlayerOrAdmin)]
         [HttpGet("groups/{groupName}/players")]
         public async Task<ActionResult> GetGroupPlayers(string groupName)
         {
@@ -479,7 +326,6 @@ namespace Nether.Web.Features.PlayerManagement
         /// <param name="group">Group name</param>
         /// <returns></returns>
         [ProducesResponseType((int)HttpStatusCode.NoContent)]
-        [Authorize(Roles = RoleNames.Admin)]
         [HttpPut("groups/{name}")]
         public async Task<ActionResult> PutGroup([FromBody]GroupPostRequestModel group)
         {
